+++ src/lm.py
@@ -41,7 +41,7 @@ class LanguageModel(nn.Module):
             pos = torch.arange(0, S).unsqueeze(0).repeat(B, 1).to(self.device)
             # x = self.dropout((x.double() @ self.tok_emb.weight * self.scale) + self.pos_emb(pos))
             x = self.dropout(((x.double() @ self.tok_emb.weight.double())* self.scale) + self.pos_emb(pos))
-            x_ = self.encoder(x, x_m, x_pm)  
+            x_ = self.encoder(x.float(), x_m, x_pm)  
             x_ = F.softmax(self.linear(x_), dim=-1)
         
         return x_
